+++ Khana POC/src/Shared/Dapp.js
@@ -4,8 +4,9 @@ import Navigation from './Navigation'
 import UserDashboard from './UserDashboard'
 import Grants from './Grants'
 import GrantHistory from './GrantHistory'
-import Admin from './Admin';
-import TokenShared from './DappTokenShared';
+import Admin from './Admin'
+import Owner from './Owner'
+import TokenShared from './DappTokenShared'
 
 class Dapp extends Component {
 
@@ -139,6 +140,17 @@ class Dapp extends Component {
                     { /* Admin section */}
                     {this.state.navigation === 3 &&
                         <Admin
+                            state={this.state}
+                            createNotification={this.createNotification}
+                            updateState={this.updateState}
+                            updateStaticState={this.updateStaticState}
+                            updateLoadingMessage={this.updateLoadingMessage}
+                        />
+                    }
+
+                    { /* Owner section */}
+                    {this.state.navigation === 4 &&
+                        <Owner
                             state={this.state}
                             createNotification={this.createNotification}
                             updateState={this.updateState}
